Файл content_template.js разделить на несколько различных модулей, выполняющих отдельные задачи
dynamic_form - сокрытие элементов формы создания и редактирования объектов при работе пользователя в этой форме

object_processing - страница отправки запросов по созданию, изменению и удалению объектов

paste_objec - фокусировка на объекте и его вставка

save_template - отправка запроса на сохранение щаблона текста

setup_queries - функции для подготовки тел post-запросов

tuple_processing - создание, изменение, удаление справочников
